Network node

+++ P2P/plugin/communication/fermat-p2p-plugin-communication-network-node-bitdubai/src/main/java/com/bitdubai/fermat_p2p_plugin/layer/communications/network/node/developer/bitdubai/version_1/structure/util/SeedServerConf.java
@@ -24,5 +24,5 @@ public class SeedServerConf {
     /**
      * Represent the DEFAULT_PORT
      */
-    public static final Integer DEFAULT_PORT = new Integer(8080);
+    public static final Integer DEFAULT_PORT = new Integer(9090);
 }
